Move test images to images/ and add a large, generated test image.
Previously we had one minimal test image.  This was located in
hivex/t (a subdirectory of the main library).

This adds a large, procedurally generated test image.  Because
this needs to be built using hivex code, and because subdirectories
are built before the parent directory by automake, we have to
also move the directory location to a top-level directory called
images/.