regedit: Add implicit nul-termination when importing strings.
When you import a string value like:
  "Foo"="Bar"
using Windows regedit program, implicit nul-termination is added
to the value (not the key), so what is stored in the value would
be something like:
  hex(1):42,00,61,00,72,00,00,00
where two of the trailing zero bytes come from the implicit
terminator.  This corrects the reg_import function so it works
the same way.