arm: Prevent cast-align warnings.
The implicit alignment of char* is 1 byte, so on ARM we get a
cast-align warning whenever we try to cast this to a larger type
(eg. casting to uint32_t).

The implicit alignment of void* is whatever malloc returns, and so
this can be cast to any type.

The cost of this is that we have to add (char *) casts all over the
code, but this seems unavoidable.

Thanks: Leif Lindholm, Peter Robinson.