ruby: tests: isolate boilerplate in common file
Isolate in a common file all the standard boilerplate in tests, i.e. the
import of the test framework and the guestfs module (including the
import path hack needed for the latter).

(This is the same change as fbae7f3e69ce598c52af71ec204f8dbd0a9ce5e5
in libguestfs.)