python: use errors more specific than RuntimeError
KeyError is documented as:
"Raised when a mapping (dictionary) key is not found in the set of
existing keys."

TypeError is documented as:
"Raised when an operation or function is applied to an object of
inappropriate type."

RuntimeError is documented as:
"Raised when an error is detected that doesn’t fall in any of the other
categories."

Let's be more specific with exceptions. The exception all inherit from
StandardError (and from Exception), but not from RuntimeError, so this
might need some code changes (if users would catch this specific error).
However, this behavior was not documented so it is perfectly fine to
change it.