Add a cache for iconv_t handles to hive_t
It was brought to my attention that dumping a registry hive causes a
lot of time spent in disk I/O activity because iconv_open() and
iconv_close() are called for every key. Every iconv_open() call causes
/usr/lib/.../gconv/$ENCODING.so to be opened and mapped.

The iconv_t handles are now cached in the hive_h struct; they are
opened on-demand and re-used.

On my ~10 year old Lenovo T60, I have seen 57% savings in the overal
runtime of running

    hivexregedit --export windows-8-enterprise-software.hive '\\'