ocaml: Use caml_alloc_initialized_string.
On OCaml 4.10 you cannot memcpy to a String_val target since
String_val returns a const char * pointer.  In any case we should be
using caml_alloc_initialized_string.  For versions of OCaml before
August 2017 which didn't have this function, provide a replacement.

The idea for this fix is copied from libnbd.