ocaml: Cast return from String_val to (char *).
This is safe here because we only pass these to functions which read
the strings (they are not freed).  A better fix would be to change
lib/hivex.h so these fields in the C struct are const char *, but that
is difficult because the library internally reuses this struct to
store some strings which it later frees.