lib: Move CPPFLAGS to AM_CPPFLAGS
When cross-compiling for mingw, there is a LIBOBJS file called mmap.c.
This must be included implicitly, but when you do this it doesn't use
the libhivex.la CPPFLAGS.  The solution seems to be to make CPPFLAGS
"global" (for this directory) by moving it to AM_CPPFLAGS so it
applies to everything compiled in this directory including the
implicit stuff.  Nothing in the automake documentation bothers to
mention this as far as I can tell.